Make the loggers truly pluggable
Remove the static list of possible loggers, and instead generate it
dynamically from the registered ones.

This requires a conceptually small but textually invasive refactoring as
to remove some undefined behavior wrt. the orders in which static
variables are initialised, we change `nix::config` from a static
variable to a function that lazily builds a `Config` object and returns
it (the reason for that is that with a static `Config`, we might − and
apparantly do in practice − get the list of registered loggers before
they are all registered, leading to an incomplete list in the
help/man/completion).